+++ scripts/preprocessing/_02_frequency_filter.py
@@ -35,7 +35,8 @@ from mne_bids import BIDSPath
 
 import config
 from config import (gen_log_kwargs, on_error, failsafe_run,
-                    import_experimental_data, import_er_data, import_rest_data)
+                    import_experimental_data, import_er_data, import_rest_data,
+                    _update_for_splits)
 from config import parallel_func
 
 
@@ -75,11 +76,10 @@ def get_input_fnames_frequency_filter(**kwargs):
 
     if cfg.use_maxwell_filter:
         bids_path_in.update(processing="sss")
-        if bids_path_in.copy().update(split='01').fpath.exists():
-            bids_path_in = bids_path_in.update(split='01')
 
     in_files = dict()
     in_files[f'raw_run-{run}'] = bids_path_in
+    _update_for_splits(in_files, f'raw_run-{run}', single=True)
 
     if (cfg.process_er or config.noise_cov == 'rest') and run == cfg.runs[0]:
         noise_task = "rest" if config.noise_cov == "rest" else "noise"
@@ -87,9 +87,8 @@ def get_input_fnames_frequency_filter(**kwargs):
             raw_noise_fname_in = bids_path_in.copy().update(
                 run=None, task=noise_task
             )
-            if raw_noise_fname_in.copy().update(split='01').fpath.exists():
-                raw_noise_fname_in.update(split='01')
             in_files["raw_noise"] = raw_noise_fname_in
+            _update_for_splits(in_files, "raw_noise", single=True)
         else:
             if config.noise_cov == 'rest':
                 in_files["raw_rest"] = bids_path_in.copy().update(
@@ -189,7 +188,7 @@ def filter_data(
 
     out_files['raw_filt'] = bids_path.copy().update(
         root=cfg.deriv_root, processing='filt', extension='.fif',
-        suffix='raw')
+        suffix='raw', split=None)
     raw.load_data()
     filter(
         raw=raw, subject=subject, session=session, run=run,
@@ -201,7 +200,9 @@ def filter_data(
     resample(raw=raw, subject=subject, session=session, run=run,
              sfreq=cfg.resample_sfreq, data_type='experimental')
 
-    raw.save(out_files['raw_filt'], overwrite=True, split_naming='bids')
+    raw.save(out_files['raw_filt'], overwrite=True, split_naming='bids',
+             split_size=cfg._raw_split_size)
+    _update_for_splits(out_files, 'raw_filt')
     if cfg.interactive:
         # Plot raw data and power spectral density.
         raw.plot(n_channels=50, butterfly=True)
@@ -238,7 +239,7 @@ def filter_data(
         out_files['raw_noise_filt'] = \
             bids_path_noise.copy().update(
                 root=cfg.deriv_root, processing='filt', extension='.fif',
-                suffix='raw')
+                suffix='raw', split=None)
 
         raw_noise.load_data()
         filter(
@@ -252,8 +253,10 @@ def filter_data(
                  sfreq=cfg.resample_sfreq, data_type=data_type)
 
         raw_noise.save(
-            out_files['raw_noise_filt'], overwrite=True, split_naming='bids'
+            out_files['raw_noise_filt'], overwrite=True, split_naming='bids',
+            split_size=cfg._raw_split_size,
         )
+        _update_for_splits(out_files, 'raw_noise_filt')
         if cfg.interactive:
             # Plot raw data and power spectral density.
             raw_noise.plot(n_channels=50, butterfly=True)
@@ -301,6 +304,7 @@ def get_config(
         min_break_duration=config.min_break_duration,
         t_break_annot_start_after_previous_event=config.t_break_annot_start_after_previous_event,  # noqa:E501
         t_break_annot_stop_before_next_event=config.t_break_annot_stop_before_next_event,  # noqa:E501
+        _raw_split_size=config._raw_split_size,
     )
     return cfg